API: move v1.0.1 specific test cases into its own package
